changed1

+++ sql_py.py
@@ -34,6 +34,7 @@ with conn.cursor() as cur:
         conn.commit()
 
     def change_person(cursor, person_id, first_name=None, last_name=None, email=None, number=None):
+
         if first_name:
             cursor.execute("""
                 UPDATE person SET first_name=%s WHERE id=%s;""",
@@ -57,40 +58,42 @@ with conn.cursor() as cur:
 
     def delete_ph_number(cursor, person_id, number):
         cursor.execute("""
-            DELETE FROM phone_number WHERE person_id=%s AND number=%s;""",
+                    SELECT * FROM phone_number WHERE person_id=%s AND number=%s;""",
+                       (person_id, number))
+        if len(cursor.fetchall()) == 0:
+            print('Такого номер нет')
+            return
+
+        cursor.execute("""
+            DELETE FROM phone_number as pn WHERE person_id=%s AND number=%s;""",
             (person_id, number))
         conn.commit()
 
     def delete_person(cursor, person_id):
         cursor.execute("""
-            DELETE FROM person WHERE id=%s;""",
+            SELECT * FROM person WHERE id=%s;""",
+            (person_id,))
+        if len(cursor.fetchall()) == 0:
+            print('Такого клиента нет')
+            return
+
+        cursor.execute("""
+            DELETE FROM person as p WHERE id=%s;""",
             (person_id,))
         conn.commit()
+        print(f'Удалена записаь о клиенте с ID {person_id}')
     def find_person(cursor, first_name=None, last_name=None, email=None, number=None):
-         if first_name:
-            cursor.execute("""
-                SELECT * FROM person WHERE first_name=%s;""",
-                (first_name,))
-            print(cur.fetchall())
-         elif last_name:
-             cursor.execute("""
-                 SELECT * FROM person WHERE last_name=%s""",
-                (last_name,))
-             print(cur.fetchall())
-         elif email:
-             cursor.execute("""
-                SELECT * FROM person WHERE email=%s""",
-                (email,))
-             print(cur.fetchall())
-         elif number:
-             cursor.execute("""
-                SELECT * FROM person WHERE id =(SELECT person_id FROM phone_number WHERE number=%s)""",
-                (number,))
-             print(cur.fetchall())
-         else:
-             cursor.execute("""
-                SELECT * FROM person""",
-                (number,))
-             print(cur.fetchall())
+         cursor.execute("""
+            SELECT p.*, pn.number FROM person as p
+            JOIN phone_number as pn ON p.id=pn.person_id
+            WHERE (first_name = %(first_name)s OR first_name = %(first_name)s is NULL)
+            AND (last_name = %(last_name)s OR last_name = %(last_name)s is NULL)
+            AND (email = %(email)s OR email = %(email)s is NULL)
+            AND (number = %(number)s OR number = %(number)s is NULL)
+            """,
+            {'first_name' : first_name, 'last_name' : last_name, 'email' : email, 'number' : number})
+         print('Найдена запись о клиенте: ', cursor.fetchall())
+
+
